WIP Unify the EEGClassifier and EEGRegressor (#501)
* WIP Unify the code

* better name

* flake8

* Fix regressor

* Update braindecode/classifier.py


* Updating the comments

* Indentation

* flake8

* trying to fix

* removing

* add name

* flake8

* Fixing the order

---------
