+++ braindecode/classifier.py
@@ -1,27 +1,21 @@
 # Authors: Maciej Sliwowski <maciek.sliwowski@gmail.com>
 #          Robin Schirrmeister <robintibor@gmail.com>
 #          Lukas Gemein <l.gemein@gmail.com>
+#          Bruno Aristimunha <b.aristimunha@gmail.com>
 #
 # License: BSD (3-clause)
 
 import warnings
 
-import numpy as np
-from sklearn.metrics import get_scorer
-from skorch.callbacks import EpochTimer, BatchScoring, PrintLog, EpochScoring
-from skorch.classifier import NeuralNet
+from skorch import NeuralNet
 from skorch.classifier import NeuralNetClassifier
-from skorch.utils import train_loss_score, valid_loss_score, noop, to_numpy
-import torch
 
-from .training.scoring import (PostEpochTrainScoring,
-                               CroppedTrialEpochScoring,
-                               CroppedTimeSeriesEpochScoring,
-                               predict_trials)
+from .eegneuralnet import _EEGNeuralNet
+from .training.scoring import predict_trials
 from .util import ThrowAwayIndexLoader, update_estimator_docstring
 
 
-class EEGClassifier(NeuralNetClassifier):
+class EEGClassifier(_EEGNeuralNet, NeuralNetClassifier):
     doc = """Classifier that does not assume softmax activation.
     Calls loss function directly without applying log or anything.
 
@@ -59,7 +53,8 @@ class EEGClassifier(NeuralNetClassifier):
     __doc__ = update_estimator_docstring(NeuralNetClassifier, doc)
 
     def __init__(self, *args, cropped=False, callbacks=None,
-                 iterator_train__shuffle=True, iterator_train__drop_last=True,
+                 iterator_train__shuffle=True,
+                 iterator_train__drop_last=True,
                  aggregate_predictions=True, **kwargs):
         self.cropped = cropped
         self.aggregate_predictions = aggregate_predictions
@@ -70,154 +65,12 @@ class EEGClassifier(NeuralNetClassifier):
                          iterator_train__drop_last=iterator_train__drop_last,
                          **kwargs)
 
-    def _yield_callbacks(self):
-        # Here we parse the callbacks supplied as strings,
-        # e.g. 'accuracy', to the callbacks skorch expects
-        for name, cb, named_by_user in super()._yield_callbacks():
-            if name == 'str':
-                train_cb, valid_cb = self._parse_str_callback(cb)
-                yield train_cb
-                if self.train_split is not None:
-                    yield valid_cb
-            else:
-                yield name, cb, named_by_user
-
-    def _parse_str_callback(self, cb_supplied_name):
-        scoring = get_scorer(cb_supplied_name)
-        scoring_name = scoring._score_func.__name__
-        assert scoring_name.endswith(
-                        ('_score', '_error', '_deviance', '_loss'))
-        if (scoring_name.endswith('_score') or
-                cb_supplied_name.startswith('neg_')):
-            lower_is_better = False
-        else:
-            lower_is_better = True
-        train_name = f'train_{cb_supplied_name}'
-        valid_name = f'valid_{cb_supplied_name}'
-        if self.cropped:
-            # TODO: use CroppedTimeSeriesEpochScoring when time series target
-            # In case of cropped decoding we are using braindecode
-            # specific scoring created for cropped decoding
-            train_scoring = CroppedTrialEpochScoring(
-                cb_supplied_name, lower_is_better, on_train=True, name=train_name
-            )
-            valid_scoring = CroppedTrialEpochScoring(
-                cb_supplied_name, lower_is_better, on_train=False, name=valid_name
-            )
-        else:
-            train_scoring = PostEpochTrainScoring(
-                cb_supplied_name, lower_is_better, name=train_name
-            )
-            valid_scoring = EpochScoring(
-                cb_supplied_name, lower_is_better, on_train=False, name=valid_name
-            )
-        named_by_user = True
-        train_valid_callbacks = [
-            (train_name, train_scoring, named_by_user),
-            (valid_name, valid_scoring, named_by_user)
-        ]
-        return train_valid_callbacks
-
-    # pylint: disable=arguments-differ
-    def get_loss(self, y_pred, y_true, *args, **kwargs):
-        """Return the loss for this batch by calling NeuralNet get_loss.
-
-        Parameters
-        ----------
-        y_pred : torch tensor
-            Predicted target values
-        y_true : torch tensor
-            True target values.
-        X : input data, compatible with skorch.dataset.Dataset
-            By default, you should be able to pass:
-
-                * numpy arrays
-                * torch tensors
-                * pandas DataFrame or Series
-                * scipy sparse CSR matrices
-                * a dictionary of the former three
-                * a list/tuple of the former three
-                * a Dataset
-
-            If this doesn't work with your data, you have to pass a
-            ``Dataset`` that can deal with the data.
-        training : bool (default=False)
-            Whether train mode should be used or not.
-
-        Returns
-        -------
-        loss : float
-            The loss value.
-        """
-        return NeuralNet.get_loss(self, y_pred, y_true, *args, **kwargs)
-
     def get_iterator(self, dataset, training=False, drop_index=True):
         iterator = super().get_iterator(dataset, training=training)
         if drop_index:
             return ThrowAwayIndexLoader(self, iterator, is_regression=False)
         else:
             return iterator
-
-    def on_batch_end(self, net, *batch, training=False, **kwargs):
-        # If training is false, assume that our loader has indices for this
-        # batch
-        if not training:
-            epoch_cbs = []
-            for name, cb in self.callbacks_:
-                if isinstance(cb, (CroppedTrialEpochScoring, CroppedTimeSeriesEpochScoring)) and (
-                        hasattr(cb, 'window_inds_')) and (not cb.on_train):
-                    epoch_cbs.append(cb)
-            # for trialwise decoding stuffs it might also be we don't have
-            # cropped loader, so no indices there
-            if len(epoch_cbs) > 0:
-                assert self._last_window_inds_ is not None
-                for cb in epoch_cbs:
-                    cb.window_inds_.append(self._last_window_inds_)
-                self._last_window_inds_ = None
-
-    def predict_with_window_inds_and_ys(self, dataset):
-        self.module.eval()
-        preds = []
-        i_window_in_trials = []
-        i_window_stops = []
-        window_ys = []
-        for X, y, i in self.get_iterator(dataset, drop_index=False):
-            i_window_in_trials.append(i[0].cpu().numpy())
-            i_window_stops.append(i[2].cpu().numpy())
-            with torch.no_grad():
-                preds.append(to_numpy(self.module.forward(X.to(self.device))))
-            window_ys.append(y.cpu().numpy())
-        preds = np.concatenate(preds)
-        i_window_in_trials = np.concatenate(i_window_in_trials)
-        i_window_stops = np.concatenate(i_window_stops)
-        window_ys = np.concatenate(window_ys)
-        return dict(
-            preds=preds, i_window_in_trials=i_window_in_trials,
-            i_window_stops=i_window_stops, window_ys=window_ys)
-
-    # Removes default EpochScoring callback computing 'accuracy' to work properly
-    # with cropped decoding.
-    @property
-    def _default_callbacks(self):
-        return [
-            ("epoch_timer", EpochTimer()),
-            (
-                "train_loss",
-                BatchScoring(
-                    train_loss_score,
-                    name="train_loss",
-                    on_train=True,
-                    target_extractor=noop,
-                ),
-            ),
-            (
-                "valid_loss",
-                BatchScoring(
-                    valid_loss_score, name="valid_loss", target_extractor=noop,
-                ),
-            ),
-            ("print_log", PrintLog()),
-        ]
 
     def predict_proba(self, X):
         """Return the output of the module's forward method as a numpy
@@ -258,10 +111,43 @@ class EEGClassifier(NeuralNetClassifier):
         # Predictions may be already averaged in CroppedTrialEpochScoring (y_pred.shape==2).
         # However, when predictions are computed outside of CroppedTrialEpochScoring
         # we have to average predictions, hence the check if len(y_pred.shape) == 3
-        if self.cropped and self.aggregate_predictions and len(y_pred.shape) == 3:
+        if self.cropped and self.aggregate_predictions and len(
+                y_pred.shape) == 3:
             return y_pred.mean(axis=-1)
         else:
             return y_pred
+
+    def get_loss(self, y_pred, y_true, *args, **kwargs):
+        """Return the loss for this batch by calling NeuralNet get_loss.
+
+        Parameters
+        ----------
+        y_pred : torch tensor
+            Predicted target values
+        y_true : torch tensor
+            True target values.
+        X : input data, compatible with skorch.dataset.Dataset
+            By default, you should be able to pass:
+
+                * numpy arrays
+                * torch tensors
+                * pandas DataFrame or Series
+                * scipy sparse CSR matrices
+                * a dictionary of the former three
+                * a list/tuple of the former three
+                * a Dataset
+
+            If this doesn't work with your data, you have to pass a
+            ``Dataset`` that can deal with the data.
+        training : bool (default=False)
+            Whether train mode should be used or not.
+
+        Returns
+        -------
+        loss : float
+            The loss value.
+        """
+        return NeuralNet.get_loss(self, y_pred, y_true, *args, **kwargs)
 
     def predict(self, X):
         """Return class labels for samples in X.
@@ -325,5 +211,6 @@ class EEGClassifier(NeuralNetClassifier):
             dataset=X,
             return_targets=return_targets,
             batch_size=self.batch_size,
-            num_workers=self.get_iterator(X, training=False).loader.num_workers,
+            num_workers=self.get_iterator(X,
+                                          training=False).loader.num_workers,
         )
